Adjust conversations tests for IRIS-1466
We can no longer expect messages "Message A" and "Message B" to be
linked into the same conversation; needless to say this broke a lot
of the conversations tests.  Use "Re: Message A" for the 2nd message.

Also, modernise some parts of the code.

Also, double-check that CIDs are different in several places where
the test implicitly depends on that behaviour.

Finally, add a new test for conversation linking which explores how much
a subject can change before a new conversation is started.