Add test for rolling squatter when mboxname locked
This was causing problems in production.  The rolling squatter could not
open a mailbox index for indexing because an imapd had the mailbox
selected and thus held the mboxname lock in shared mode.  Simulate this
by turning off the code in replication_wait which carefully disconnects
and reconnects the clients.